+++ lib/recording_studio_support/engine.rb
@@ -137,3 +137,5 @@ module RecordingStudioSupport
     end
   end
 end
+
+require_relative "engine/javascript"