+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_09_09_023312) do
+ActiveRecord::Schema[8.1].define(version: 2026_09_09_080000) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_catalog.plpgsql"
   enable_extension "pgcrypto"
@@ -187,6 +187,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_09_09_023312) do
 
   create_table "recording_studio_support_sections", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.datetime "created_at", null: false
+    t.string "icon"
     t.string "slug", null: false
     t.string "title", null: false
     t.index ["slug"], name: "index_rs_support_sections_on_slug"